Changed calculator and test
Found an issue in my latest fix that wasn't really fixed. What was happening was if the new interval date was a Thursday then it would add my interval that may have put the date on a Saturday or Sunday which wasn't checked for in the code, I added those checks so weekend days would be skipped over.